Adição da Lógica do Valor a Receber

+++ Api3Dsm/src/main/java/com/api3Dsm/domain/controle/ClienteControle.java
@@ -57,8 +57,8 @@ public class ClienteControle {
 	@PostMapping("/inserir")
 	@ResponseStatus(HttpStatus.CREATED)
 	public void cadastrarCliente(@Valid @RequestBody Cliente cliente){
-		Double precoDoServico = cliente.getServico().getPreco();
-        Double precoCadaParcela = precoDoServico / 12;
+		float precoDoServico = cliente.getServico().getPreco();
+        float precoCadaParcela = precoDoServico / 12;
         List<Parcela> listaParcelas = new ArrayList<>();
 		LocalDate dataVenci = LocalDate.now();
